refactor: switch to ballerina (#6)
The lsp (gtache) plugin we based ours on crashed often when a flux source
was invalid (failing type checks) and then, while in that state, the editor
asks the LSP for completion or signature info.

This diff switches us to depend on the "ballerina" LSP library which is
effectively a fork of the original plugin we tried to use.

"Ballerina" still has issues in similar situations, but is a little more
recoverable. If the editor gets in a bad state and stops offering
completions, closing and opening the buffer again will usually get it
going again.

While code reformatting didn't work at all with the gtache plugin, it
ALSO doesn't work here with ballerina. It appears the support is there,
but the config (in plugin.xml) may be outdated for the current intellij
platform.

---

N.b. the instructions for configuring the lib in the current master
branch README are incorrect. I had to refer to the README from the
0.94.2 tag to get this all wired up.