wip: partial grammar
Working my way through the various leaf items. So far, the following seem
to work (though some are cheating a bit):

- identifiers
- int literals
- float literals
- duration literals
- date time literals
- string literals (big cheating here)
- regex literals (same cheating as string lit)

Learning that the regex rules used to define tokens are difficult to
control. They are greedy, so overlapping patterns are not easy to
define or prioritize.

Rather than use regex, we should probably be defining more complex rules
in a ParserUtil class (which I have stubbed). It might even be possible
to leverage our rust parser that way... TBD.

String literals are where this issue showed up. Defining a producer for
`hex_digit` clashed with `identifier` which currently relies on `letter`
(a regex) since `letter` _has to be regex_ to match the broad unicode
`Letter` category.

My hope is delegating these sorts of checks to a util class will allow
us to match these cases more selectively rather than GLOBALLY as is the
case with the regex token definitions.