wip: `Expression` as common root, allowing for left recursion
Looks like the solution for pipe forward, etc involves making all the
various types of expression _extend_ a common root.

To support this, ValueExpr was removed as a distinct sub-rule. The root
is now favored instead. `IdentExpr` was added to fold `identity` in with
`QualifiedExpr` to cover one and two part names (ie `pkg.ident`).